refactor: Phase 1.2 - consolidate dependency files into unified module
Consolidate 3 dependency files into single dependency/ module:
- dependency.py (221 LOC) → deleted
- service_layer/dependency.py (61 LOC) → deleted
- models/dependency.py (38 LOC) → moved to dependency/models.py

New structure:
- dependency/service.py (331 LOC) - unified implementation
- dependency/models.py (38 LOC) - data models
- dependency/__init__.py (33 LOC) - public API

Changes:
- Merged build_dependency_graph(), to_dot(), DependencyService
- Enhanced documentation and type hints
- Updated all imports (5 files)

Results:
- LOC: 14,456 → 14,571 (+115 for better docs)
- Better organization and documentation
- All functionality preserved

Refs: v1.8.0 Phase 1.2
