+++ src/snowcli_tools/service_layer/__init__.py
@@ -1,7 +1,7 @@
 """Service layer abstractions for snowcli-tools."""
 
 from ..catalog import CatalogService
-from .dependency import DependencyService
+from ..dependency import DependencyService
 from .query import QueryService
 
 __all__ = ["CatalogService", "DependencyService", "QueryService"]
